Always resolve realpaths when comparing paths
This fixes cases where the gem is loaded from a symlinked
folder, and we end up with the file path not being under
the gem full path location. We also do a realpath lookup for the source
location path too, so that if the source location is in a symlinked
folder, for example in tests, we don't break.