Extract `ActionController::Helpers` (#60)
* extract action-controller

* wip

* test decorate

* wip tests

* wip

* Lookup instance methods on the constant first

Looking up methods on the helper class has a problem since Rails creates
proxy methods (with the same name) for some helper methods on the helper
module. This messes up the generated parameters and prevents us from
ever getting the correct signature of the actual method.

Instead, if we look for the instance method on the `constant` first,
then we can return the method that we found for processing, since
ultimately we know that the parameters and the parameter types will that
of the method on the constant.

If the method does not exist on the constant, we fallback to looking up
the instance method on the helper module.

This has a slight chance of generating the wrong output if a helper
method defined in a `helper {}` and a method defined on the controller
end up having the same name but are totally unrelated. To prevent that
we could check for the source location of the helper instance method and
if it matches the location in `abstract_controller/helpers.rb` where
`helper_method` proxies are created, only then access the method on the
controller.

* Make sure we only access method on constant if it is proxied

Rails stores the list of method names that it has setup proxies for in
the `_helper_methods` attribute, so instead of always trying to match
methods on the constant by name, we can check if the method name is in
that list and if so access the method on the controller instead of the
proxy on the helper module.

This solves the problem pointed out in the previous commit about the
edge case where blindly looking up by name might lead to generating the
wrong output.

* add sig to block syntax test

* Fix problem with void return signatures

* update docs

Co-authored-by: Ufuk Kayserilioglu <ufuk.kayserilioglu@shopify.com>