Start searching parent paths for gemspec in git gems
This is necessary for the cases where the location of the loaded file
could end up being different from the gem path indicated by the spec
file.

The solution is to walk up the parent folder hierarchy for a given file,
until the desired gemspec file is matched. This is only performed for
git gems.