Make all `belongs_to` associations return nilable
Active Record model objects exhibit different type behaviour in different
contexts. For an Active Record model that is created in memory, the
attributes and associations can all be `nil`. The strict typing on the
object properties are only applied when a model is read from the
database.

The signature generation code was looking at the database column
nilability to understand if the association property returns a nilable
type or not. It turns out that that is not the correct way to represent
the association method return types, since this is totally valid Ruby on
Rails code:

```ruby
class Model < ActiveRecord::Base
  belongs_to :foo
end

model = Model.new
model.foo #=> nil
```

even when the database column that the `foo` association maps to is
strictly not nilable.

Thus, the solution is to make all associations to return a nilable type
by default.