Fix parameter handling in signatures
Fixes #116

There were a couple of problems with parameter handling in signatures:

1. We were messing up the parameter ordering in the signature which
sometimes ended up not matching with the order in which parameters were
declared in the method parameter list. This is fixed by iterating over
the actual method parameters and only using signature parameter mapping
as a lookup table to get the relevant type for the parameter.
2. We were skipping `keyrest` arguments completely. :facepalm:

I also had to keep a sanitized version of the parameter list to pass to
the signature compile step, since we end up modifying the parameter
names while generating the method definition. The parameter name in the
signature should match that changed name, so the sanitized parameter
list should be used when processing the signatures.