Fix ActiveRecordColumns spec StrongTypeGeneration handling
Contrary to what the documentation of `ActiveRecordColumns` generator
says, we actually generate strong types for models if
`StrongTypeGeneration` is not defined in the user application as a
constant.

We only generate weak types if a constant named `StrongTypeGeneration`
is defined in the user application and the current model does not extend
from it.

Added separate test cases that test default behaviour and behaviour when
`StrongTypeGeneration` is defined.