Error early and clear files for unprocessable constant names
After we try to convert requested constant names to constants, we check
to see if there were anuy failures. If there were, we display an error
for each failed constant lookup and we try to delete the associated RBI
file if it exists.

We also exit with a non-zero exit code.