Mixed in class methods detection should be best effort
We try to do something sneaky to detect modules that are mixed into a
class through inclusion of another module, where we try to infer that
information from the change in the ancestor list of a test class after
including the said module to that class.

However, that operation is risky since a `self.included` method on the
target module could potentially run any Ruby code and it turns out some
code in the wild even try to require optional gems when the concern is
being included.

Since this detection is best effort, it should be fine to catch
all kinds of `Exception`s and bail out of the operation without
side-effects.

Fixes: #237