Ensure safety against subclass prevention methods
Previously, we were doing special casing to ensure that we were not
caught by the `inherited` hook in a sealed class preventing us from
subclassing the generic type. However, that was short-sighted since it
only worked for sealed classes. But there are more cases where
subclassing is not allowed and they are all handled via `inherited`
hooks.

So, the generic (hah, nice pun), way of addressing this is to override
the `inherited` hook as we are creating a subclass and then placing the
method back in its place. That makes sure that no-one can throw because
we are subclassing.