Fix `FixedHash` serialization
`T::Types::FixedHash` did not used to differentiate between the key
types and serialized all keys as Symbols. This created an information
loss when Tapioca was serializing parameters in signatures that had
fixed hash values with String or mixed keys.

The same change is done upstream in Sorbet in https://github.com/sorbet/sorbet/pull/4223
and this is a backport of that change to older versions of Sorbet.