Add ActiveSupport::Concern DSL generator
Sorbet currently does not provide a way of indicating that a module
mixes in class methods from its ancestors. Through this RBI generator
we can figure out which of its ancestors contain a `ClassMethods`
module and provide that information to Sorbet using
`mixes_in_class_methods`.

In order to pull this off, we first identify all named modules in the
`ObjectSpace` that are not singleton classes, that extend
`ActiveSupport::Concern` and have dependencies (i.e. include other
concerns).

For those modules, we iterate over their linearized dependencies and
pick out all modules that define a `ClassMethods` constant.

Finally we append these `ClassMethods` modules in a
`mixes_in_class_methods` call in the correct order.