Check existence before looking up proxy target method
In order to handle cases where a controller has not defined the target
of a helper method proxy, we should check to see if the target method is
defined before trying to look it up on the constant.

If the proxy target method does not exist, then we should define a
generic, untyped proxy method.