Stop monkey-patching Class to add descendants
This causes a confusion with the `Class#descendants` method that Active
Support adds and the signature on this method leaks out into the Tapioca
generated RBI and causes type-checking problems in Core because of it.

This is arguably a reflection method, so I moved it to
`Tapioca::Reflection` module which is already included in the Base DSL
class and SymbolGenerator class.