Start filtering mixins by matching locations to gem path
We should generate a mixin in the RBI file only if we can be sure that
the mixin location was coming from a gem file or if we could not track
the mixin location at all.

In all other cases, we can safely drop the mixin, since it should be
generated in the RBI file of the gem that is actually performing the
mixin in the first place.

Co-authored-by: Alexandre Terrasa <583144+Morriar@users.noreply.github.com>