Start patching `append_features`/`prepend_features`/`extend_object`
Since some code checks or relies on the result of `callers` or
`caller_locations` during a mixin, if we patch mixin methods, we might
get spurious failures.

A safer way is to patch `append_features` to catch `include`s,
`prepend_features` to catch `prepend`s and `extend_object`s to catch
`extend`s.

Co-authored-by: Alan Wu <XrXr@users.noreply.github.com>