Stop intercepting `Module#autoload` with a prepend
We need to do the alias-method-chain dance since Bootsnap does the same,
and prepended modules and alias-method-chain don't play well together.

So, why does Bootsnap do alias-method-chain and not prepend? Glad you
asked! That's because RubyGems does alias-method-chain for
`Kernel#require` and such, so, if Bootsnap were to do `prepend`, it
might end up breaking RubyGems.