Trim require path prefixes from default gem filenames
In Ruby 3.1, the filenames in `spec.files` of default gems have
a require path prefix, even though relative to the Ruby lib dir, that
prefix path does not exist. For example, for `ostruct`, the only file
is listed as `lib/ostruct.rb` instead of `ostruct.rb`.

Thus, when constructing the files list, we need to trim those require
path prefixes, if they exist, from filenames, before we look for it
relative to the Ruby lib dir.