Revert back to using local references to type variables
Previous changes made references to type variables fully qualified,
however, for Sorbet it is valid for a type to reference another type's
type variables, so fully qualified references were only creating a
distraction.

By changing the constant name extraction method from a custom name to
being a `name` override (and special casing that in the `T::Types` name
handler), we are able to refer to these references locally.