All tests should use `spec_helper` for unified setup
Some tests were doing their own setup inside `spec_with_project` but
that meant that any Minitest configuration would have to be duplicated.

However, `spec_with_project` just defines a `Minitest::Spec` subclass,
so it can easily be required from `spec_helper` as part of the setup
for all tests.

This fixes a problem with CLI tests that we just relying on requiring
`spec_with_project` not respecting the line number filter for the
runner, since the configuration of Minitest was different there.