Coerce `TypeVariableModule` instances to custom type
Sorbet runtime checks for method arguments change depending on if the
method arguments are all simple types and/or the return type is a simple
type or not.

Ever since we introduced the `TypeVariableModule` to make name binding
easier, type members/templates started being wrapped in simple types
when coerced, as well. However, this ends up breaking the assumptions
that Sorbet runtime is making, thus causing test failures for DSL
compilers with generic typed arguments in the constructor.

The solution is to hook the `T::Utils.coerce` method to coerce those
types to a custom `T::TypeVariable` subclass that records its name. By
doing this, we preserve the checks that Sorbet runtime is doing and we
preserve the name of the type variable.