Fix type variable patch to respect references to late defined types
Now that Sorbet syntax for declaring type variables is via a block,
code authors can very well refer from that block to types that have not
been defined yet. Since the values of `fixed`, `lower` and `upper` are
late evaluated, this should work normally.

However, our current attempt to retrofit the new type variable syntax
was trying to early evaluate the block that was passed so that we could
unify on the old syntax. However, that early evaluation breaks for the
case outlined above.

We should instead convert the bounds passed to type variables into the
block format, if a block is not passed, and always work with blocks
until the point where we need to serialize the type variable. We should
be pretty guaranteed that the types referred to from the block will be
defined by the time we get to serialize the type variable. In any case,
that is as late we can evaluate the bounds anyway.