Restore Gemfile and Gemfile.lock before each test
For speed reasons, the project is only setup once, but if a gem is
added, it can leak to other tests.

This saves Gemfile and Gemfile.lock before all `generate` tests, and
restores it before each of the test, which avoids re-creating the whole
project, or even just running `bundle install` which takes a couple of
seconds.