Add `--environment` flag to `dsl`/`gem` commands
Tapioca is primarily a development tool, and thus, should be used to
load applications/gems in `development` mode by default. Up to now,
however, we assumed that we could just be not opinionated about this,
which led to mismatch between the output of the commands locally and on
CI system (which usually run in `test` environment).

By adding an `--environment` flag that defaults to `development`, we
can allow users to use the sensible default, but also to override it.