Fix method name validation edge-case
It turns out that `:4_to_5=` is not a valid symbol but `"4_to_5=".to_sym`
does not quote it and represents it as `:4_to_5=`, which throws off our
whole method name validation logic.

The fix is to check validity of the name after stripping trailing `=`
as well. That leaves `==` as an edge-case, that we handle separately.