Make `ConstantDefinition` tracker use `resolve_loc`
This allows dynamically created constants to be attributed to the
file that triggered their creation instead of the actual location that
created the module.

For example, given a file `foo.rb`:
```ruby
module Foo
  class_methods do
    # ...
  end
end
```
we expect a `Foo::ClassMethods` module to be created when this file is
processed. That module will be created inside Active Support using
something like `const_set(:ClassMethods, Module.new)`, but that module
should be attributed to `foo.rb` where the `class_methods` call is
being made.