Fix no_toc annotation
The Kramdown block attribute was not working properly on Github and we
ended up showing a `{:.no_doc}` on the README page.

This commit fixes that by moving that annotation to an HTML comment on
the raw text of the header that we match against to filter it out.

I also took the opportunity to encapsulate some of the operations we
were performing into a custom converter class that inherits from the
built-in `Toc` converter. This makes the code more readable.