Avoid `exit` call in the command classes
Thor already knows how to deal with `exit` status.

If call `exit` ourselves, it become harder to reuse those commands,
in say, profiling calls, because as soon the command suceeds, we were
exiting the entire process, not allowing the profiler to dump the
data.

If we use the mechamism inside Thor to handle exits, we also can
simplify the stderr output code.

This also allowed us to fix some TODOs that were in the tests about
the error not being sent to stderr.