Add flag to list compilers for DSL command
Since we are now loading all compilers from any loaded gem that
exports a DSL compiler, the full set of DSL compilers in play might be
a little confusing.

Thus, this PR adds a way to ask Tapioca to list the set of DSL compilers
and get a tabular output that shows them as well as their enabled/disabled
status.