Don't type required_ancestor's block
The reason it takes block is to prevent the content inside it from being
evaluated. And the block's content is actually not relevant in the runtime.

Tapioca is the rare or even only exception where we do use the block's
return value in an unexpected way. And since this usage (hack) is not really
supported by Sorbet, we should make it clear by using `T.untyped` instead.