Start matching pathname directory prefixes more rigorously
Fixes #1174

The simplistic `start_with?` check was not sufficient to match that a
given file/folder was under a given folder, since we were not explicitly
checking for full folder name matches. For example, if
`path = "/foo/bar-gem/baz"` and `folder = "/foo/bar"`, then
`path.start_with?(folder)` would return `true`, but `path` is not under
the `folder` directory.

This commit fixes the problem by matching `folder` to any of the parent
directories of `path` by using the `Pathname#ascend` method to get all
parent directory prefixes of a `path`.