Add failing test for additions to just the singleton class of existing classes
Sorbet does not include the base class as a symbol in the symbol table
if the source file only contains singleton class operations on types
that exist in the payload.

Since we skip singleton classes in the symbol table parser, we never
consider the target class for such additions for RBI generation.